refactor: introduce doInTransaction helper method
