+++ query-node/substrate-query-framework/index-builder/src/db/helper.ts
@@ -1,5 +1,5 @@
 import * as shortid from 'shortid';
-import { DeepPartial, createConnection, Connection, getConnectionOptions } from 'typeorm';
+import { DeepPartial, createConnection, QueryRunner, Connection, getConnection, getConnectionOptions } from 'typeorm';
 
 import { SnakeNamingStrategy } from './SnakeNamingStrategy';
 /**
@@ -27,4 +27,27 @@ export function fillRequiredWarthogFields<T>(entity: DeepPartial<T>): DeepPartia
 export async function createDBConnection(): Promise<Connection> {
   const connectionOptions = await getConnectionOptions();
   return createConnection(Object.assign(connectionOptions, { namingStrategy: new SnakeNamingStrategy() }));
+}
+
+
+export async function doInTransaction(fn: (qn: QueryRunner) => Promise<void>): Promise<void> {
+  const queryRunner = getConnection().createQueryRunner();
+  try {
+    // establish real database connection
+    await queryRunner.connect();
+    await queryRunner.startTransaction();
+    
+    await fn(queryRunner);
+
+    await queryRunner.commitTransaction();
+  } catch (error) {
+    console.error(`Rolling back the transaction due to errors: ${JSON.stringify(error, null, 2)}`);
+
+    // Since we have errors lets rollback changes we made
+    await queryRunner.rollbackTransaction();
+    throw new Error(error);
+  } finally {
+    // Query runner needs to be released manually.
+    await queryRunner.release();
+  }
 }
